fastapi_app/requirements.txt
pytest
black
requests
pyjwt
python-dotenv
apispec
pylint
fastapi
uvicorn
sqlalchemy
mysqlclient
httpx
python-multipart